+++ Models/Prestamos.cs
@@ -1,3 +1,4 @@
+
 using System.ComponentModel.DataAnnotations;
 
 public class Prestamos
@@ -5,14 +6,12 @@ public class Prestamos
         [Key]
     public int PrestamosId {get; set;}
     
-
     [Required (ErrorMessage ="La descripción es requerida")]
 
     public DateTime? fecha{get; set;}
     public DateTime? vence{get; set;}
     public string? concepto{get;set;}
-    public double? monto{get;set;}
+    public long? monto{get;set;}
     public int? personaid{get;set;}
     public int? balance {get; set;}
-
 }
